Splittig of functions
